.migration-backup/backend/requirements.txt
django>=6.0.4
django-cors-headers>=4.9.0
django-filter>=25.2
djangorestframework>=3.17.1
djangorestframework-simplejwt>=5.5.1
dj-database-url>=2.2.0
google-auth>=2.38.0
psycopg[binary]>=3.2.9
python-dotenv>=1.0.1
requests>=2.32.0
reportlab>=4.4.10
whitenoise>=6.6.0
gunicorn>=22.0.0
